# Reshard fusion with all_gather_concat + model integration