--- a/models/demos/llama3_subdevices/tt/llama_attention.py
+++ b/models/demos/llama3_subdevices/tt/llama_attention.py
@@ -369,14 +369,7 @@ class TtLlamaAttention(LightweightModule):
         # attn_output_1G4D_sharded.deallocate(True)
 
         # Note: Persistent output buffer used, do not deallocate output!
-        attn_output_gathered_sharded = self.tt_ccl.line_all_gather(
-            attn_output_1G4D_sharded,
-            dim=1,
-            cluster_axis=1,
-            num_links=3,
-            memory_config=self.model_config["GATHER_USERS_MEMCFG"](list(self.mesh_device.shape)[1]),
-            buffer_key="SDPA",
-        )
+
         # ttnn.deallocate(attn_output_1G4D)
 
         # attn_output_gathered_sharded = ttnn.to_memory_config(
@@ -384,8 +377,23 @@ class TtLlamaAttention(LightweightModule):
         # )
         # ttnn.deallocate(attn_output_gathered)
 
-        attn_output_cat_0 = ttnn.experimental.nlp_concat_heads_decode(
-            attn_output_gathered_sharded,
+        output_shard_shape = (32, 128)
+        output_shard_grid = ttnn.CoreRangeSet(
+            {
+                ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(3, 1)),
+                ttnn.CoreRange(ttnn.CoreCoord(1, 2), ttnn.CoreCoord(2, 2)),
+            }
+        )
+        output_shard_spec = ttnn.ShardSpec(output_shard_grid, output_shard_shape, ttnn.ShardOrientation.ROW_MAJOR)
+        output_mem_config = ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.WIDTH_SHARDED, buffer_type=ttnn.BufferType.L1, shard_spec=output_shard_spec
+        )
+        attn_output_cat_0 = self.tt_ccl.all_gather_concat(
+            attn_output_1G4D_sharded,
+            dim=1,
+            cluster_axis=1,
+            num_links=3,
+            memory_config=output_mem_config,
             num_heads=self.n_local_heads,
         )
         # print("done concat heads")