--- a/models/demos/llama3_subdevices/tt/llama_attention.py
+++ b/models/demos/llama3_subdevices/tt/llama_attention.py
@@ -372,34 +372,24 @@ class TtLlamaAttention(LightweightModule):
 
         # ttnn.deallocate(attn_output_1G4D)
 
-        output_shard_shape = (32, 128)
-        output_shard_grid = ttnn.CoreRangeSet(
-            {
-                ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(3, 1)),
-                ttnn.CoreRange(ttnn.CoreCoord(1, 2), ttnn.CoreCoord(2, 2)),
-            }
-        )
-        output_shard_spec = ttnn.ShardSpec(output_shard_grid, output_shard_shape, ttnn.ShardOrientation.ROW_MAJOR)
-        output_mem_config = ttnn.MemoryConfig(
-            ttnn.TensorMemoryLayout.WIDTH_SHARDED, buffer_type=ttnn.BufferType.L1, shard_spec=output_shard_spec
-        )
-        attn_output_cat_0 = self.tt_ccl.all_gather_concat(
+        # attn_output_gathered_sharded = ttnn.to_memory_config(
+        #     attn_output_gathered, self.model_config["GATHER_USERS_MEMCFG"](list(self.mesh_device.shape)[1])
+        # )
+        # ttnn.deallocate(attn_output_gathered)
+
+        attn_output_cat = self.tt_ccl.all_gather_concat(
             attn_output_1G4D_sharded,
             dim=1,
             cluster_axis=1,
             num_links=3,
-            memory_config=output_mem_config,
+            memory_config=self.model_config["SHARDED_ATTN_WO_INPUT_RING_MEMCFG"],
             num_heads=self.n_local_heads,
         )
         # print("done concat heads")
 
         # Original matmul on each device [1, 1, 32, 1024] @ [1, 1, 1024, 2048]
-        # attn_output_cat = ttnn.to_memory_config(
-        #    attn_output_cat_0, self.model_config["SHARDED_ATTN_WO_INPUT_RING_MEMCFG"]
-        # )
-        # attn_output_cat_0.deallocate(True)
         dense_out_ttnn = ttnn.matmul(
-            attn_output_cat_0,
+            attn_output_cat,
             self.wo,
             program_config=self.model_config["WO_DECODE_RING_PROGCFG"],
             memory_config=self.model_config["SHARDED_WO_OUT_RING_MEMCFG"],
@@ -409,7 +399,6 @@ class TtLlamaAttention(LightweightModule):
             sub_device_id=self.prefetcher_setup.worker_sub_device_id,
         )
         # [1, 1, 32, 2304]
-        # ttnn.deallocate(attn_output_cat)
         # print("done matmul")
 
         dense_out_reduced = self.tt_ccl.line_all_reduce(