--- a/models/demos/llama3_subdevices/tt/llama_ccl.py
+++ b/models/demos/llama3_subdevices/tt/llama_ccl.py
@@ -29,6 +29,7 @@ class TT_CCL:
         self.num_cbs = 2
         self.from_remote_semaphore_handles = []
         self.to_remote_semaphore_handles = []
+        self.all_gather_concat_inter_tensor = self.get_all_gather_concat_inter_buffer()
 
         # Double buffered on each axis
         self.gather_semaphore_handles = [[], []]
@@ -64,6 +65,36 @@ class TT_CCL:
         self.gather_idx = [0, 0]
         self.buffer_idx = [0, 0]
         self.reduce_scatter_buffer_idx = [0, 0]
+
+    def get_all_gather_concat_inter_buffer(self):
+        intermediate_core_range_set = ttnn.CoreRangeSet(
+            {
+                ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(3, 9)),
+                ttnn.CoreRange(ttnn.CoreCoord(5, 0), ttnn.CoreCoord(6, 0)),
+            }
+        )
+        intermediate_mem_config = ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.HEIGHT_SHARDED,
+            ttnn.BufferType.L1,
+            ttnn.ShardSpec(
+                intermediate_core_range_set,
+                [32, 128],
+                ttnn.ShardOrientation.ROW_MAJOR,
+            ),
+        )
+        temp_shape = [8, 128, 32, 128]
+        intermediate_tensor = torch.zeros(temp_shape, dtype=torch.bfloat16)
+        tt_intermediate_tensor = ttnn.from_torch(
+            intermediate_tensor,
+            device=self.mesh_device,
+            layout=ttnn.TILE_LAYOUT,
+            dtype=ttnn.bfloat16,
+            memory_config=intermediate_mem_config,
+            mesh_mapper=ttnn.ShardTensor2dMesh(self.mesh_device, dims=[0, 1], mesh_shape=[8, 4]),
+        )
+        check_mesh_tensor_alloc(tt_intermediate_tensor)
+        tt_intermediate_tensors = [tt_intermediate_tensor]
+        return tt_intermediate_tensors
 
     def get_all_gather_buffers(self):
         """
@@ -530,6 +561,24 @@ class TT_CCL:
         )
         self.gather_idx[cluster_axis] = (self.gather_idx[cluster_axis] + 1) % self.num_cbs
         # ttnn.synchronize_device(self.mesh_device, sub_device_ids=[self.worker_sub_device_id])
+        return ttnn_tensor_out
+
+    def all_gather_concat(self, input_tensor_mesh, dim, cluster_axis, memory_config, num_links=1, num_heads=8):
+        ttnn_tensor_out = ttnn.experimental.all_gather_concat(
+            input_tensor_mesh,
+            self.all_gather_concat_inter_tensor[0],
+            dim,
+            cluster_axis=cluster_axis,
+            mesh_device=self.mesh_device,
+            topology=ttnn.Topology.Linear,
+            multi_device_global_semaphore=self.gather_semaphore_handles[cluster_axis][self.gather_idx[cluster_axis]],
+            num_links=num_links,
+            num_heads=num_heads,
+            memory_config=memory_config,
+            subdevice_id=self.worker_sub_device_id,
+            enable_persistent_fabric_mode=self.enable_persistent_fabric,
+        )
+        self.gather_idx[cluster_axis] = (self.gather_idx[cluster_axis] + 1) % self.num_cbs
         return ttnn_tensor_out
 
     def line_all_reduce_host(self, input_tensor_mesh, cluster_axis, num_links, memory_config):