--- a/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat.cpp
+++ b/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat.cpp
@@ -19,8 +19,8 @@ ttnn::Tensor ExecuteAllGatherConcat::invoke(
     const MeshDevice& mesh_device,
     const global_semaphore::MultiDeviceGlobalSemaphore& multi_device_global_semaphore,
     const uint32_t num_heads,
+    const ttnn::MemoryConfig& memory_config,
     const std::optional<uint32_t> num_links,
-    const std::optional<ttnn::MemoryConfig>& memory_config,
     const ttnn::ccl::Topology topology,
     std::optional<tt::tt_metal::SubDeviceId> subdevice_id) {
     return ttnn::operations::experimental::ccl::all_gather_concat(
@@ -31,8 +31,8 @@ ttnn::Tensor ExecuteAllGatherConcat::invoke(
         mesh_device,
         multi_device_global_semaphore,
         num_heads,
-        num_links,
         memory_config,
+        num_links,
         topology,
         subdevice_id);
 }