--- a/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat.hpp
+++ b/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat.hpp
@@ -21,8 +21,8 @@ struct ExecuteAllGatherConcat {
         const MeshDevice& mesh_device,
         const global_semaphore::MultiDeviceGlobalSemaphore& multi_device_global_semaphore,
         const uint32_t num_heads,
+        const ttnn::MemoryConfig& memory_config,
         const std::optional<uint32_t> num_links = std::nullopt,
-        const std::optional<ttnn::MemoryConfig>& memory_config = std::nullopt,
         const ttnn::ccl::Topology topology = ttnn::ccl::Topology::Linear,
         std::optional<tt::tt_metal::SubDeviceId> subdevice_id = std::nullopt);
 };