--- a/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat_pybind.cpp
+++ b/ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_concat_heads_fused/all_gather_concat_pybind.cpp
@@ -34,8 +34,8 @@ void bind_all_gather_concat(pybind11::module& module, const ccl_operation_t& ope
                const MeshDevice& mesh_device,
                const global_semaphore::MultiDeviceGlobalSemaphore& multi_device_global_semaphore,
                const uint32_t num_heads,
+               const ttnn::MemoryConfig& memory_config,
                const std::optional<uint32_t> num_links,
-               const std::optional<ttnn::MemoryConfig>& memory_config,
                const ttnn::ccl::Topology topology,
                std::optional<tt::tt_metal::SubDeviceId> subdevice_id,
                QueueId queue_id) -> ttnn::Tensor {
@@ -48,8 +48,8 @@ void bind_all_gather_concat(pybind11::module& module, const ccl_operation_t& ope
                     mesh_device,
                     multi_device_global_semaphore,
                     num_heads,
-                    num_links,
                     memory_config,
+                    num_links,
                     topology,
                     subdevice_id);
             },
@@ -60,9 +60,9 @@ void bind_all_gather_concat(pybind11::module& module, const ccl_operation_t& ope
             py::arg("mesh_device"),
             py::arg("multi_device_global_semaphore"),
             py::arg("num_heads").noconvert(),
+            py::arg("memory_config"),
             py::kw_only(),
             py::arg("num_links") = 1,
-            py::arg("memory_config") = std::nullopt,
             py::arg("topology") = ttnn::ccl::Topology::Linear,
             py::arg("subdevice_id") = std::nullopt,
             py::arg("queue_id") = DefaultQueueId});