### Ticket
Llama 80tkps project

### Problem description
Fused all_gather_concat with the following reshard operation
Integrated the op in the llama model

The kernel time dropped from 19.8 us in main to 16.9 us
The op to op latency dropped from 1.95 us to 0.79 us
The non overlapped dispatch time dropped from 15.73 us to 11.46 us

### What's changed
- Changed the specs, intermediate cores, output cores and NLP concat worker cores to align with the reshard output
- Separate the all_gather and the NLP worker cores: it is now 3 for all_gather and 16 for NLP concat
- Replaced all_gather, nlp_concat_heads and reshard ops with the fused op in the llama_attention

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes