@ntarafdar 2025-04-21
hey Nour can you verify if the observed savings matches what we laid out and put the savings in the PR description